LayoutTests/http/tests/notifications/notification-in-non-secure-context-expected.txt

CONSOLE MESSAGE: The Notification permission may only be requested in a secure context.
CONSOLE MESSAGE: The Notification permission may only be requested in a secure context.

PASS Notification permission in insecure context
